refactor: isolate ai loop
