+++ .github/workflows/update_digest.yml
@@ -9,32 +9,13 @@ jobs:
   build-and-push:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v3
-        with:
-          persist-credentials: true
-
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-
-      - name: Generate Trending Digest
-        run: python -m src.render_digest
-
-      - name: Run tests
-        run: pytest tests/test_trending.py
-
-      - name: Commit and push if changed
+      - uses: actions/checkout@v3
+      - run: pip install -r requirements.txt
+      - run: python -m src.render_digest
+      - name: Commit updated files
         run: |
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
           git add TRENDING.md README.md
-          git commit -m "chore: update trending digest [skip ci]" || echo "No changes to commit"
-          git pull --rebase origin main --autostash
-          git push origin main
+          git commit -m "ci: update trending digest"
+          git push
