+++ ai_loop/scripts/setup_wizard.py
