+++ ai_loop/tests/test_codex_autobot.py
@@ -4,7 +4,9 @@ from pathlib import Path
 
 import importlib.util
 
-MODULE_PATH = Path(__file__).resolve().parents[1] / 'ai_loop' / 'codex_autobot.py'
+os.environ.setdefault('OPENAI_API_KEY', 'test')
+
+MODULE_PATH = Path(__file__).resolve().parents[2] / 'ai_loop' / 'codex_autobot.py'
 spec = importlib.util.spec_from_file_location('autobot', MODULE_PATH)
 autobot = importlib.util.module_from_spec(spec)
 spec.loader.exec_module(autobot)
